split backend/common test
